fix: folder renamed and services folder relocated in applications
