+++ src/contexts/SystemHealth/application/useCases/CheckDBHealthStatus.ts
@@ -1,9 +1,9 @@
 import { DBHealth } from "../../domain/implementations/DBHealth";
-import prismaClient from "../../../Shared/infrastrucutre/database/prismaClient";
+import prismaClient from "@/contexts/Shared/infrastructure/database/PrismaClient";
 
 export class CheckDBHealthUseCase {
   private prisma = prismaClient;
-  constructor() {}
+  constructor() { }
 
   async execute(): Promise<DBHealth> {
     try {
